@@ -101,33 +101,49 @@
     "eslint-plugin-react-hooks": "4.6.0",
     "fs-extra": "^10.1.0",
     "husky": "^8.0.3",
-    "jest": "^29.4.1",
-    "jest-environment-jsdom": "^29.4.1",
-    "jsonc-eslint-parser": "^2.2.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jsonc-eslint-parser": "^2.4.0",
     "kill-port": "^2.0.1",
-    "lint-staged": "^13.0.3",
+    "lint-staged": "^13.3.0",
     "memfs": "3.4.13",
-    "node-fetch": "3.3.0",
+    "minimatch": "^9.0.5",
+    "node-fetch": "3.3.2",
     "nx": "19.5.1",
-    "openapi-types": "^12.1.0",
-    "prettier": "^3.0.3",
-    "prettier-2": "npm:prettier@^2",
+    "openapi-types": "^12.1.3",
+    "prettier": "^3.3.3",
+    "prettier-2": "npm:prettier@^2.8.8",
     "react-refresh": "^0.10.0",
-    "run-p": "*",
+    "run-p": "^0.0.0",
     "semantic-release": "^19.0.5",
-    "style-loader": "^3.3.0",
+    "style-loader": "^3.3.4",
     "stylus": "0.59.0",
-    "stylus-loader": "^7.1.0",
+    "stylus-loader": "^7.1.3",
     "tree-kill": "^1.2.2",
     "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
-    "tslib": "^2.5.0",
+    "tslib": "^2.6.3",
     "typescript": "5.5.3",
     "url-loader": "^4.1.1",
-    "verdaccio": "^5.24.1",
-    "webpack": "^5.81.0",
-    "webpack-merge": "^5.8.0",
+    "verdaccio": "^5.31.1",
+    "webpack": "^5.93.0",
+    "webpack-merge": "^5.10.0",
     "yargs": "^17.7.2"
+  },
+  "resolutions": {
+    "@nx/devkit": "19.5.1",
+    "@nx/cypress": "19.5.1",
+    "@nx/eslint": "19.5.1",
+    "@nx/eslint-plugin": "19.5.1",
+    "@nx/jest": "19.5.1",
+    "@nx/js": "19.5.1",
+    "@nx/node": "19.5.1",
+    "@nx/plugin": "19.5.1",
+    "@nx/react": "19.5.1",
+    "@nx/web": "19.5.1",
+    "@nx/webpack": "19.5.1",
+    "@nx/workspace": "19.5.1",
+    "nx": "19.5.1"
   },
   "repository": {
     "type": "git",
